+++ include/checker.h
@@ -6,7 +6,7 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/14 10:26:54 by reasuke           #+#    #+#             */
-/*   Updated: 2024/02/14 10:29:44 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/19 12:11:08 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,5 +15,7 @@
 
 # include "../libft/libft.h"
 # include "types.h"
+
+void	sort_based_on_operation(t_stack **p_a, t_stack **p_b);
 
 #endif
